Move dateTime utils into src
